Update project.html
In this update, I played with some colors of the buttons. I changed the save button to next so that instead of going back to the main menu, the next button will save the input and go to the program's next page. It also saves the all inputs and shows them on the summary page.